requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
jinja2==3.1.5
httpx==0.28.1
yt-dlp==2024.12.23
python-telegram-bot==21.9
google-api-python-client==2.159.0
google-auth-oauthlib==1.2.1
pydantic==2.10.4
pydantic-settings==2.7.1
aiofiles==24.1.0
python-dotenv==1.0.1
openai==1.58.1
